Asynchronous compile and back pressure
Split compile/1 into two parts: submit/1 and pickup/2.

submit/1 returns immediately, either with a job reference or `busy`.
This enables callers to do other work whilst waiting and to back off
when the server is busy.

pickup/2 returns immediately if there are results ready or waits for
a job to complete. The caller supplies the wait timeout.